Added style file for flex layout
